refactor(tts): extract directives and providers to separate modules
Split tts.ts (1585→1026 lines) into three focused modules:
- tts-directives.ts (392 lines): tag parsing, validation, normalization
- tts-providers.ts (249 lines): ElevenLabs, OpenAI, Edge implementations
- tts.ts (1026 lines): orchestration, config resolution, main API

All 34 tests pass. No public API changes.

lobster-biscuit
